refactor: move helper test function to common file
